% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial.R
\name{calc_resid}
\alias{calc_resid}
\title{Calculate surrogate residuals for multinomial/discrete choice models}
\usage{
calc_resid(y, proba)
}
\arguments{
\item{y}{Vector of observed classes (either a factor or a 1-based integer vector).}

\item{proba}{Matrix of predicted probabilities, where rows correspond to
observations and columns correspond to classes.}
}
\value{
A matrix of surrogate residuals of the same dimensions as \code{proba}.
}
\description{
Calculate surrogate residuals for multinomial logit and general discrete choice models
using a fast, exact conditional Gumbel simulation method.
}
