% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.heteroscedastic}
\alias{df.heteroscedastic}
\title{Simulated Heteroscedastic Data}
\format{A data frame with 2000 rows and 2 variables.
\itemize{
  \item \code{x} The predictor variable.
  \item \code{y} The response variable; an ordered factor.
}}
\usage{
data(df.heteroscedastic)
}
\description{
Data simulated from a probit model with heteroscedasticity.
}
\examples{
head(df.heteroscedastic)
}
\keyword{datasets}
