% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.quadratic}
\alias{df.quadratic}
\title{Simulated Quadratic Data}
\format{A data frame with 2000 rows and 2 variables.
\itemize{
  \item \code{x} The predictor variable.
  \item \code{y} The response variable; an ordered factor.
}}
\usage{
data(df.quadratic)
}
\description{
Data simulated from a probit model with a quadratic trend.
}
\examples{
head(df.quadratic)
}
\keyword{datasets}
