% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df2}
\alias{df2}
\title{Simulated Heteroscedastic Data}
\format{A data frame with 2000 rows and 2 variables.
\itemize{
  \item \code{x} The predictor variable.
  \item \code{y} The response variable; an ordered factor.
}}
\usage{
data(df2)
}
\description{
Data simulated from a probit model with heteroscedasticity. The data are
described in Example 4 of Liu and Zhang (2017).
}
\examples{
head(df2)
}
\references{
Liu, Dungang and Zhang, Heping. Residuals and Diagnostics for Ordinal
Regression Models: A Surrogate Approach.
\emph{Journal of the American Statistical Association} (accepted).
}
\keyword{datasets}
