% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df5}
\alias{df5}
\title{Simulated Interaction Data}
\format{A data frame with 2000 rows and 3 variables.
\itemize{
  \item \code{y} The response variable; an ordered factor.
  \item \code{x1} A continuous predictor.
  \item \code{x2} A factor with two levels: \code{Control} and
  \code{Treatment}.
}}
\usage{
data(df5)
}
\description{
Data simulated from an ordered probit model with an interaction term.
}
\examples{
head(df5)
}
\keyword{datasets}
