% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getFittedProbs}
\alias{getFittedProbs}
\title{#' @keywords internal
getFittedValues <- function(object) {
  UseMethod("getFittedValues")
}}
\usage{
getFittedProbs(object)
}
\description{
#' @keywords internal
getFittedValues <- function(object) {
  UseMethod("getFittedValues")
}
}
\keyword{internal}
