% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{gof}
\alias{gof}
\alias{gof.clm}
\alias{gof.polr}
\alias{gof.vglm}
\alias{plot.gof}
\title{Kolmogorov-Smirnov Simulation}
\usage{
gof(object, nsim = 10, test = c("ks", "ad", "cvm"), ...)

\method{gof}{clm}(object, nsim = 10, test = c("ks", "ad", "cvm"), ...)

\method{gof}{polr}(object, nsim = 10, test = c("ks", "ad", "cvm"), ...)

\method{gof}{vglm}(object, nsim = 10, test = c("ks", "ad", "cvm"), ...)

\method{plot}{gof}(x, ...)
}
\arguments{
\item{object}{An object of class \code{\link[ordinal]{clm}},
\code{\link[MASS]{polr}}, or \code{\link[VGAM]{vglm}}.}

\item{nsim}{Integer specifying the number of bootstrap replicates to use.}

\item{test}{Cahracter string specifying which goodness-of-fit test to use.
Options include: \code{"ks"} for the Kolmogorov-Smirnov test; \code{"ad"} for
the Anderson-Darling test, and \code{"cvm"} for the Cramer-Von Mises test.
Default is \code{"ks"}.}

\item{...}{Additional optional arguments. (Currently ignored.)}

\item{x}{An object of class \code{"gof"}.}
}
\value{
An object of class \code{c("gof", "numeric")}.
}
\description{
Simulate p-values from a goodness-of-fit test (experimental).
}
\details{
Under the null hypothesis, the distribution of the p-values should appear
uniformly distributed on the interval [0, 1]. This can be visually
investigated using the \code{plot} method. A 45 degree line is indicative of
a "good" fit.
}
\examples{
\dontrun{
# Load required packages
library(MASS)
library(ordr)

# Fit a proportional odds model
house.polr <- polr(Sat ~ Infl + Type + Cont, weights = Freq, data = housing,
                   method = "logistic")

# Goodness-of-fit
plot(gof(house.polr, nsim = 1000))
}
}
