% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksTest.R
\name{ksTest}
\alias{ksTest}
\alias{ksTest.clm}
\alias{ksTest.polr}
\alias{ksTest.vglm}
\alias{plot.ksTest}
\title{Kolmogorov-Smirnov Simulation}
\usage{
ksTest(object, nsim = 10, ...)

\method{ksTest}{clm}(object, nsim = 10, ...)

\method{ksTest}{polr}(object, nsim = 10, ...)

\method{ksTest}{vglm}(object, nsim = 10, ...)

\method{plot}{ksTest}(x, ...)
}
\arguments{
\item{object}{An object of class \code{\link[ordinal]{clm}},
\code{\link[MASS]{polr}}, or \code{\link[VGAM]{vglm}}.}

\item{nsim}{Integer specifying the number of bootstrap replicates to use.}

\item{...}{Additional optional arguments. (Currently ignored.)}

\item{x}{An object of class \code{"ksTest"}.}
}
\value{
An object of class \code{c("ksTest", "numeric")}.
}
\description{
Simulate p-values from the Kolmogorov-Smirnov test (experimental).
}
