% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.resid}
\alias{plot.resid}
\alias{plot.clm}
\alias{plot.glm}
\alias{plot.lrm}
\alias{plot.orm}
\alias{plot.polr}
\alias{plot.vglm}
\title{Plot Residual Diagnostics}
\usage{
\method{plot}{resid}(
  x,
  y = NULL,
  what = c("qq", "fitted", "covariate"),
  covariate = NULL,
  fit = NULL,
  distribution = qnorm,
  ncol = NULL,
  alpha = 1,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  smooth = TRUE,
  ...
)

\method{plot}{clm}(
  x,
  y = NULL,
  what = c("qq", "fitted", "covariate"),
  covariate = NULL,
  fit = NULL,
  distribution = qnorm,
  ncol = NULL,
  alpha = 1,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  smooth = TRUE,
  ...
)

\method{plot}{glm}(
  x,
  y = NULL,
  what = c("qq", "fitted", "covariate"),
  covariate = NULL,
  fit = NULL,
  distribution = qnorm,
  ncol = NULL,
  alpha = 1,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  smooth = TRUE,
  ...
)

\method{plot}{lrm}(
  x,
  y = NULL,
  what = c("qq", "fitted", "covariate"),
  covariate = NULL,
  fit = NULL,
  distribution = qnorm,
  ncol = NULL,
  alpha = 1,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  smooth = TRUE,
  ...
)

\method{plot}{orm}(
  x,
  y = NULL,
  what = c("qq", "fitted", "covariate"),
  covariate = NULL,
  fit = NULL,
  distribution = qnorm,
  ncol = NULL,
  alpha = 1,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  smooth = TRUE,
  ...
)

\method{plot}{polr}(
  x,
  y = NULL,
  what = c("qq", "fitted", "covariate"),
  covariate = NULL,
  fit = NULL,
  distribution = qnorm,
  ncol = NULL,
  alpha = 1,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  smooth = TRUE,
  ...
)

\method{plot}{vglm}(
  x,
  y = NULL,
  what = c("qq", "fitted", "covariate"),
  covariate = NULL,
  fit = NULL,
  distribution = qnorm,
  ncol = NULL,
  alpha = 1,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  smooth = TRUE,
  ...
)
}
\arguments{
\item{x}{An object of class \code{"resid"} (returned by \code{\link[=resids]{resids()}}),
or a fitted model object of class \code{\link[ordinal:clm]{ordinal::clm()}},
\code{\link[stats:glm]{stats::glm()}}, \code{\link[rms:lrm]{rms::lrm()}}, \code{\link[rms:orm]{rms::orm()}},
\code{\link[MASS:polr]{MASS::polr()}}, or \code{\link[VGAM:vglm]{VGAM::vglm()}}.}

\item{y}{Ignored.}

\item{what}{Character string specifying what to plot. Default is \code{"qq"}
which produces a quantile-quantile plot of the residuals.}

\item{covariate}{A vector giving the covariate values to use for residual-by-
covariate plots (i.e., when \code{what = "covariate"}).}

\item{fit}{The fitted model from which the residuals were extracted. (Only
required if \code{what = "fitted"} and \code{x} inherits from class
\code{"resid"}.)}

\item{distribution}{Function that computes the quantiles for the reference
distribution to use in the quantile-quantile plot. Default is \code{qnorm}
which is only appropriate for models using a probit link function. When
\code{jitter_scale = "probability"}, the reference distribution is always
U(-0.5, 0.5).}

\item{ncol}{Integer specifying the number of columns to use for the plot
layout (if requesting multiple plots). Default is \code{NULL}.}

\item{alpha}{A single value in the interval \verb{[0, 1]} controlling the
opacity alpha of the plotted points. Only used when \code{nsim} > 1.}

\item{xlab}{Character string giving the text to use for the x-axis label.
Default is \code{NULL}.}

\item{ylab}{Character string giving the text to use for the y-axis label.
Default is \code{NULL}.}

\item{main}{Character string giving the plot title. Default is \code{NULL}.}

\item{smooth}{Logical indicating whether or not to add a nonparametric
smooth to certain plots. Default is \code{TRUE}.}

\item{...}{Additional optional arguments to be passed onto
\code{\link[tinyplot:tinyplot]{tinyplot::tinyplot()}} (e.g., \code{col}, \code{pch}, \code{cex},
etc.) or \code{\link[=resids]{resids()}} (if calling the plot method directly on a
fitted model object).}
}
\value{
No return value, called for plotting.
}
\description{
Residual-based diagnostic plots for cumulative link and general regression
models using \code{\link[tinyplot:tinyplot]{tinyplot::tinyplot()}} graphics.
}
\note{
For \code{vglm} (and \code{vgam}) objects, the \strong{VGAM} package defines its own S4 \code{plot}
method which will override this S3 method if \strong{VGAM} is loaded. In such cases,
we recommend extracting the residuals first using \code{resids()} and then plotting the
resulting \code{resid} object (e.g. \code{plot(resids(fit))}), which will correctly dispatch
to \code{plot.resid()}.
}
