% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sure-package.R
\docType{package}
\name{sure-package}
\alias{sure}
\alias{sure-package}
\title{sure: An R package for constructing surrogate-based residuals and diagnostics
for ordinal and general regression models.}
\description{
The \code{sure} package provides surrogate-based residuals for fitted ordinal
and general (e.g., binary) regression models of class
\code{\link[ordinal:clm]{ordinal::clm()}}, \code{\link[stats:glm]{stats::glm()}}, \code{\link[rms:lrm]{rms::lrm()}},
\code{\link[rms:orm]{rms::orm()}}, \code{\link[MASS:polr]{MASS::polr()}}, or
\code{\link[VGAM:vglm]{VGAM::vglm()}}.
}
\details{
The development version can be found on GitHub:
\url{https://github.com/bgreenwell/sure}. As of right now, \code{sure} exports the
following functions:
\itemize{
\item \code{resids()} - construct (surrogate-based) residuals;
\item \code{plot()} - plot diagnostics using \code{\link[tinyplot:tinyplot]{tinyplot::tinyplot()}}-based graphics;
\item \code{gof()} - simulate p-values from a goodness-of-fit test.
}
}
\references{
Liu, Dungang and Zhang, Heping. Residuals and Diagnostics for Ordinal
Regression Models: A Surrogate Approach.
\emph{Journal of the American Statistical Association} (accepted).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bgreenwell/sure}
  \item \url{https://bgreenwell.github.io/sure/}
  \item \url{https://bgreenwell.r-universe.dev/sure}
  \item Report bugs at \url{https://github.com/bgreenwell/sure/issues}
}

}
\author{
\strong{Maintainer}: Brandon Greenwell \email{greenwell.brandon@gmail.com}

Authors:
\itemize{
  \item Andrew McCarthy \email{mccarthyan@gmail.com}
  \item Boehmke Bradley \email{bradleyboehmke@gmail.com}
}

Other contributors:
\itemize{
  \item Liz Glaser \email{liz.glaser@gmail.com} [contributor]
}

}
\keyword{internal}
