% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.resid}
\alias{autoplot.resid}
\alias{autoplot.clm}
\alias{autoplot.polr}
\alias{autoplot.vglm}
\title{Residual Plots for Ordinal Regression Models}
\usage{
autoplot.resid(object, what = c("qq", "mean", "covariate"), x = NULL,
  fit = NULL, distribution = qnorm, alpha = 1, xlab = NULL,
  color = "#444444", qqpoint.color = "#444444", qqpoint.shape = 19,
  qqpoint.size = 2, qqline.color = "#888888", qqline.linetype = "dashed",
  qqline.size = 1, smooth = TRUE, smooth.color = "#0000FF",
  smooth.linetype = 1, smooth.size = 1, fill = NULL, ...)

autoplot.clm(object, what = c("qq", "mean", "covariate"), x = NULL,
  nsim = 1, alpha = 1, xlab = NULL, ylab = NULL, color = "#444444",
  qqpoint.color = "#444444", qqpoint.shape = 19, qqpoint.size = 2,
  qqline.color = "#888888", qqline.linetype = "dashed", qqline.size = 1,
  smooth = TRUE, smooth.color = "#0000FF", smooth.linetype = 1,
  smooth.size = 1, fill = NULL, ...)

autoplot.polr(object, what = c("qq", "mean", "covariate"), x = NULL,
  nsim = 1, alpha = 1, xlab = NULL, ylab = NULL, color = "#444444",
  qqpoint.color = "#444444", qqpoint.shape = 19, qqpoint.size = 2,
  qqline.color = "#888888", qqline.linetype = "dashed", qqline.size = 1,
  smooth = TRUE, smooth.color = "#0000FF", smooth.linetype = 1,
  smooth.size = 1, fill = NULL, ...)

autoplot.vglm(object, what = c("qq", "mean", "covariate"), x = NULL,
  nsim = 1, alpha = 1, xlab = NULL, ylab = NULL, color = "#444444",
  qqpoint.color = "#444444", qqpoint.shape = 19, qqpoint.size = 2,
  qqline.color = "#888888", qqline.linetype = "dashed", qqline.size = 1,
  smooth = TRUE, smooth.color = "#0000FF", smooth.linetype = 1,
  smooth.size = 1, fill = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{\link[ordinal]{clm}},
\code{\link[MASS]{polr}}, \code{\link[VGAM]{vglm}}, or \code{"resid"}.}

\item{what}{Character string specifying what to plot. Default is \code{"qq"}
which produces a quantile-quantile plots of the residuals.}

\item{x}{A vector giving the covariate values to use for residual-by-
covariate plots (i.e., when \code{what = "covariate"}).}

\item{fit}{The fitted model. Only required when \code{object} is of class
\code{"resid"} and \code{what = "mean"}.}

\item{distribution}{Function that computes the quantiles for the reference
distribution to use in the quantile-quantile plot. Default is \code{qnorm}
which is only appropriate for models using a probit link function.}

\item{alpha}{A single values in the interval [0, 1] controlling the opacity
alpha of the plotted points. Only used when \code{nsim} > 1.}

\item{xlab}{Character string giving the text to use for the x-axis label in
residual-by-covariate plots. Default is \code{NULL}.}

\item{smooth}{Logical indicating whether or not too add a nonparametric
smooth on certain plots. Default is \code{TRUE}.}

\item{fill}{Character string or integer specifying the color to use to fill
the boxplots for residual-by-covariate plots when \code{x} is of class
\code{"factor"}. Default is \code{NULL} which colors the boxplots according
to the factor levels.}

\item{...}{Additional optional arguments. (Currently ignored.)}

\item{nsim}{Integer specifying the number of bootstrap replicates to use.}

\item{ylab}{Character string giving the text to use for the y-axis label.
Default is \code{NULL}.}

\item{main}{Character string giving the text to use for the plot title.
Default is \code{NULL}.}

\item{col}{Character string or integer specifying what color to use for the
points in the quantile-quantile and residual-by-fitted value/covariate plots.
Defaul is \code{"black"}.}

\item{smooth.col}{Character string or integer specifying what color to use
for the nonparametric smooth. Default is \code{"blue"}.}
}
\value{
A \code{"ggplot"} object.
}
\description{
Residual-based diagnostic plots for ordnial regression models using
\code{\link[ggplot2]{ggplot2}} graphics.
}
