% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df4}
\alias{df4}
\title{Simulated proportionality data}
\format{A data frame with 4000 rows and 2 variables.
\itemize{
  \item \code{x} The predictor variable.
  \item \code{y} The response variable; an ordered factor.
}}
\usage{
data(df4)
}
\description{
Data simulated from from two separate probit models. The data are described
in Example 5 of Liu and Zhang (2017).
}
\examples{
head(df4)
}
\references{
Liu, Dungang and Zhang, Heping. Residuals and Diagnostics for Ordinal
Regression Models: A Surrogate Approach.
\emph{Journal of the American Statistical Association} (accepted).
}
\keyword{datasets}
