% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordr.R
\docType{package}
\name{ordr}
\alias{ordr}
\alias{ordr-package}
\title{ordr: An R package for constructing residuals and diagnostics for ordinal
regression models.}
\description{
The \code{ordr} package provides surrogate and jittered residuals for fitted
ordinal regression models of class \code{\link[ordinal]{"clm"}},
\code{\link[MASS]{"polr"}}, and \code{\link[VGAM]{"vglm"}}.
}
\references{
Liu, Dungang and Zhang, Heping. Residuals and Diagnostics for Ordinal
Regression Models: A Surrogate Approach.
\emph{Journal of the American Statistical Association} (accepted).
}
