% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resids.R
\name{resids}
\alias{resids}
\alias{resids.default}
\alias{resids.glm}
\title{Residuals Cumulative Link and General Regression Models}
\usage{
resids(object, ...)

\method{resids}{default}(object, nsim = 1L, ...)

\method{resids}{glm}(object, jitter.scale = c("response", "probability"),
  nsim = 1L, ...)
}
\arguments{
\item{object}{An object of class \code{\link[ordinal]{clm}},
\code{\link[stats]{glm}}, \code{\link[MASS]{polr}}, or
\code{\link[VGAM]{vglm}}.}

\item{...}{Additional optional arguments. (Currently ignored.)}

\item{nsim}{Integer specifying the number of bootstrap replicates to use.
Default is \code{1L} meaning no bootstrap samples.}

\item{jitter.scale}{Character string specifying the scale on which to perform
the jittering. Should be one of \code{"response"} or \code{"probability"}.
Currently only used when object inherits from class \code{"glm"}. Default is
\code{"response"}.}
}
\value{
A numeric vector (\code{nsim = 1}) or matrix (\code{nsim} > 1) of
residuals. If \code{nsim} > 1, then the result will be a matrix with
\code{nsim} columns, one for each bootstrap repliacte of the residuals. The
result will contain the additional class \code{"resid"}, which is recognized
by \code{\link{autoplot.resid}}.
}
\description{
Surrogate-based residuals for cumulative link and general regression models.
}
\note{
The internal functions used for sampling from truncated distirbutions
are based on modified versions of \code{\link[truncdist]{rtrunc}} and
\code{\link[truncdist]{qtrunc}}.
}
\references{
Liu, Dungang and Zhang, Heping. Residuals and Diagnostics for Ordinal
Regression Models: A Surrogate Approach.
\emph{Journal of the American Statistical Association} (accepted). URL
http://www.tandfonline.com/doi/abs/10.1080/01621459.2017.1292915?journalCode=uasa20

Nadarajah, Saralees and Kotz, Samuel. R Programs for Truncated Distributions.
\emph{Journal of Statistical Software, Code Snippet}, 16(2), 1-8, 2006. URL
https://www.jstatsoft.org/v016/c02.
}
