% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resids.R
\name{resids}
\alias{resids}
\alias{resids.default}
\title{Residuals for Ordinal Regression Models}
\usage{
resids(object, ...)

\method{resids}{default}(object, type = c("surrogate", "jitter"),
  jitter.scale = c("response", "probability"), nsim = 1, ...)
}
\arguments{
\item{object}{An object of class \code{\link[ordinal]{clm}},
\code{\link[MASS]{polr}}, or \code{\link[VGAM]{vglm}}.}

\item{...}{Additional optional arguments. (Currently ignored.)}

\item{type}{Character string specifying the type of residuals to construct.
Should be one of \code{"surrogate"} or \code{"jitter"}. Default is
\code{"surrogate"}. (Currently ignored.)}

\item{jitter.scale}{Character string specifying the scale on which to perform
the jittering. Should be one of \code{"response"} or \code{"probability"}.
Only used when \code{type = "jitter"}. Default is \code{"response"}.
(Currently ignored.)}

\item{nsim}{Integer specifying the number of bootstrap replicates to use.}
}
\value{
A numeric vector (\code{nsim = 1}) or matrix (\code{nsim} > 1) of
residuals. If \code{nsim} > 1, then the result will be a matrix with
\code{nsim} columns, one for each bootstrap repliacte of the residuals.
}
\description{
Surrogate and jittered residuals for ordinal regression models.
}
\references{
Liu, Dungang and Zhang, Heping. Residuals and Diagnostics for Ordinal
Regression Models: A Surrogate Approach.
\emph{Journal of the American Statistical Association} (accepted).
}
