% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resids.R
\name{resids}
\alias{resids}
\alias{resids.default}
\title{Residuals for Ordinal Regression Models}
\usage{
resids(object, ...)

\method{resids}{default}(object, type = c("surrogate", "jitter"), nsim = 1,
  ...)
}
\arguments{
\item{object}{}

\item{type}{}

\item{nsim}{Integer specifying the number of bootstrap replicates to use.}
}
\value{
A numeric vector of residuals.
}
\description{
Surrogate and jittered residuals for ordinal regression models.
}
\references{
Liu, Dungang and Zhang, Heping. Residuals and Diagnostics for Ordinal
Regression Models: A Surrogate Approach.
\emph{Journal of the American Statistical Association} (accepted).
}
