% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resplot.R
\name{resplot}
\alias{resplot}
\alias{resplot.resid}
\alias{resplot.vglm}
\title{Residual Plots for Ordinal Regression Models}
\usage{
resplot(object, what = c("qq", "covariate"), x = NULL, ...)

\method{resplot}{resid}(object, what = c("qq", "covariate"), x = NULL,
  distribution = qnorm, alpha = 1, xlab = NULL, ylab = NULL,
  main = NULL, ...)

\method{resplot}{vglm}(object, what = c("qq", "covariate"), x = NULL,
  nsim = 1, alpha = 1, xlab = NULL, ylab = NULL, main = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{\link[ordinal]{clm}},
\code{\link[MASS]{polr}}, or \code{\link[VGAM]{vglm}}.}

\item{what}{Character string specifying what to plot. Default is \code{"qq"}
which produces a quantile-quantile plots of the residuals.}

\item{x}{A vector giving the covariate values to use for residual-by-
covariate plots (i.e., when \code{what = "covariate"}).}

\item{...}{Additional optional arguments. (Currently ignored.)}

\item{distribution}{Function that computes the quantiles for the reference
distribution to use in the quantile-quantile plot. Default is \code{qnorm}
which is only appropriate for models using a probit link function.}

\item{alpha}{A single values in the interval [0, 1] controlling the opacity
alpha of the plotted points. Only used when \code{nsim} > 1.}

\item{xlab}{Character string giving the text to use for the x-axis label.
Default is \code{NULL}}

\item{ylab}{Character string giving the text to use for the y-axis label.}

\item{main}{Character string giving the text to use for the plot title.}

\item{nsim}{Integer specifying the number of bootstrap replicates to use.}
}
\description{
Residual-based diagnostic plots for ordnial regression models.
}
