% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sure.R
\docType{package}
\name{sure}
\alias{sure}
\title{sure: An R package for constructing surrogate-based residuals and diagnostics
for ordinal and general regression models.}
\description{
The \code{sure} package provides surrogate-based residuals for fitted ordinal
and general (e.g., binary) regression models of class
\code{\link[ordinal]{clm}}, \code{\link[stats]{glm}}, \code{\link[rms]{lrm}},
\code{\link[rms]{orm}}, \code{\link[MASS]{polr}}, or
\code{\link[VGAM]{vglm}}.
}
\details{
The development version can be found on GitHub:
\url{https://github.com/AFIT-R/sure}. As of right now, \code{sure} exports the
following functions:
\itemize{
  \item{\code{resids}} - construct (surrogate-based) residuals;
  \item{\code{autoplot}} - plot diagnostics using
  \code{\link[ggplot2]{ggplot2}}-based graphics;
  \item{\code{gof}} - simulate p-values from a goodness-of-fit test.
}
}
\references{
Liu, Dungang and Zhang, Heping. Residuals and Diagnostics for Ordinal
Regression Models: A Surrogate Approach.
\emph{Journal of the American Statistical Association} (accepted).
}
