% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sure.R
\docType{package}
\name{sure}
\alias{sure}
\alias{sure-package}
\title{sure: An R package for constructing surrogate-based residuals and diagnostics
for ordinal and general regression models.}
\description{
The \code{sure} package provides surrogate-based residuals for fitted ordinal
and general (e.g., binary) regression models of class
\code{\link[ordinal]{clm}}, \code{\link[stats]{glm}},
\code{\link[MASS]{polr}}, and \code{\link[VGAM]{vglm}}.
}
\details{
The development version can be found on GitHub:
https://github.com/AFIT-R/sure. As of right now, \code{sure} only exports two
functions:
\itemize{
  \item{\code{resids}} - construct (surrogate-based) residuals for
  \code{\link[ordinal]{clm}}, \code{\link[stats]{glm}},
  \code{\link[MASS]{polr}}, and \code{\link[VGAM]{vglm}} objects;
  \item{\code{autoplot}} - plot diagnostics for \code{\link[ordinal]{clm}},
\code{\link[stats]{glm}}, \code{\link[MASS]{polr}}, and
\code{\link[VGAM]{vglm}} objects using \code{\link[ggplot2]{ggplot2}}
graphics;
  \item{\code{gof}} - simulate p-values from a goodness-of-fit test.
}
}
\references{
Liu, Dungang and Zhang, Heping. Residuals and Diagnostics for Ordinal
Regression Models: A Surrogate Approach.
\emph{Journal of the American Statistical Association} (accepted).
}
